+++ src/reactGameKit/utils/gameLoop.js
